requirements.txt
hvac==0.11.0
Jinja2==3.0.1
PyYAML==5.4.1
requests==2.26.0
schema==0.7.4
python-gitlab==2.10.1